requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
gunicorn>=21.2
whitenoise>=6.6
psycopg2-binary>=2.9
python-dotenv>=1.0
stripe>=5.0.0
Pillow>=10.0.0
requests>=2.32.0
